+++ src/algorithms.js
@@ -1,12 +1,12 @@
 // Start algorithm class
 let ASTRO = require('./astro');
-let ON = require('./on');
+let State = require('./on');
 
 class Algorithms {
     constructor (parent) {
         this.parent = parent;
         this.ASTRO = new ASTRO();
-        this.ON = new ON();
+        this.State = new State();
         /*  You may notice that a variety of array variables logically local
          to functions are declared globally here.  In JavaScript, construction
          of an array variable from source code occurs as the code is
@@ -364,87 +364,87 @@ class Algorithms {
         let j, year, mon, mday, hour, min, sec,
           weekday, utime, perscal;
 
-        year = this.ON.gregorian.year;
-        mon = this.ON.gregorian.month;
-        mday = this.ON.gregorian.day;
-        hour = 0;//this.ON.gregorian.hour;
-        min = 0;//this.ON.gregorian.minute;
-        sec = 0;//this.ON.gregorian.second;
+        year = this.State.gregorian.year;
+        mon = this.State.gregorian.month;
+        mday = this.State.gregorian.day;
+        hour = 0;//this.State.gregorian.hour;
+        min = 0;//this.State.gregorian.minute;
+        sec = 0;//this.State.gregorian.second;
 
-        this.ON.gDate = new Date(
+        this.State.gDate = new Date(
           year,
           mon,
           mday,
-          this.ON.gregorian.hour,
-          this.ON.gregorian.minute,
-          this.ON.gregorian.second,
-          this.ON.gregorian.millisecond
+          this.State.gregorian.hour,
+          this.State.gregorian.minute,
+          this.State.gregorian.second,
+          this.State.gregorian.millisecond
         );
 
         if (this.parent._utcMode === false) {
-            this.ON.zone = this.ON.gDate.getTimezoneOffset();
+            this.State.zone = this.State.gDate.getTimezoneOffset();
         }
 
 
         // Added for this algorithms cant parse 2016,13,32 successfully
-        this.ON.gregorian.year = this.ON.gDate.getFullYear();
-        this.ON.gregorian.month = this.ON.gDate.getMonth();
-        this.ON.gregorian.day = this.ON.gDate.getDate();
+        this.State.gregorian.year = this.State.gDate.getFullYear();
+        this.State.gregorian.month = this.State.gDate.getMonth();
+        this.State.gregorian.day = this.State.gDate.getDate();
 
         //  Update Julian day
         // ---------------------------------------------------------------------------
         j = this.gregorian_to_jd(year, mon + 1, mday) +
           (Math.floor(sec + 60 * (min + 60 * hour) + 0.5) / 86400.0);
 
-        this.ON.julianday = j;
-        this.ON.modifiedjulianday = j - this.JMJD;
+        this.State.julianday = j;
+        this.State.modifiedjulianday = j - this.JMJD;
 
         //  Update day of week in Gregorian box
         // ---------------------------------------------------------------------------
         weekday = this.ASTRO.jwday(j);
         // Move to 1 indexed number
-        this.ON.gregorian.weekday = weekday + 1;
+        this.State.gregorian.weekday = weekday + 1;
 
         //  Update leap year status in Gregorian box
         // ---------------------------------------------------------------------------
-        this.ON.gregorian.leap = this.NormLeap[this.leap_gregorian(year) ? 1 : 0];
+        this.State.gregorian.leap = this.NormLeap[this.leap_gregorian(year) ? 1 : 0];
 
         //  Update Julian Calendar
         // ---------------------------------------------------------------------------
 //        julcal = this.jd_to_julian(j);
 //
-//        this.ON.juliancalendar.year = julcal[0];
-//        this.ON.juliancalendar.month = julcal[1] - 1;
-//        this.ON.juliancalendar.day = julcal[2];
-//        this.ON.juliancalendar.leap = this.NormLeap[this.leap_julian(julcal[0]) ? 1 : 0];
+//        this.State.juliancalendar.year = julcal[0];
+//        this.State.juliancalendar.month = julcal[1] - 1;
+//        this.State.juliancalendar.day = julcal[2];
+//        this.State.juliancalendar.leap = this.NormLeap[this.leap_julian(julcal[0]) ? 1 : 0];
         weekday = this.ASTRO.jwday(j);
-//        this.ON.juliancalendar.weekday = weekday;
+//        this.State.juliancalendar.weekday = weekday;
 
         //  Update Persian Calendar
         // ---------------------------------------------------------------------------
         if (this.parent.calendarType == 'persian' && this.parent.leapYearMode == 'algorithmic') {
             perscal = this.jd_to_persian(j);
-            this.ON.persian.year = perscal[0];
-            this.ON.persian.month = perscal[1] - 1;
-            this.ON.persian.day = perscal[2];
-            this.ON.persian.weekday = this.gWeekDayToPersian(weekday);
-            this.ON.persian.leap = this.NormLeap[this.leap_persian(perscal[0]) ? 1 : 0];
+            this.State.persian.year = perscal[0];
+            this.State.persian.month = perscal[1] - 1;
+            this.State.persian.day = perscal[2];
+            this.State.persian.weekday = this.gWeekDayToPersian(weekday);
+            this.State.persian.leap = this.NormLeap[this.leap_persian(perscal[0]) ? 1 : 0];
         }
 
         //  Update Persian Astronomical Calendar
         // ---------------------------------------------------------------------------
         if (this.parent.calendarType == 'persian' && this.parent.leapYearMode == 'astronomical') {
             perscal = this.jd_to_persiana(j);
-            this.ON.persianAstro.year = perscal[0];
-            this.ON.persianAstro.month = perscal[1] - 1;
-            this.ON.persianAstro.day = perscal[2];
-            this.ON.persianAstro.weekday = this.gWeekDayToPersian(weekday);
-            this.ON.persianAstro.leap = this.NormLeap[this.leap_persiana(perscal[0]) ? 1 : 0];
+            this.State.persianAstro.year = perscal[0];
+            this.State.persianAstro.month = perscal[1] - 1;
+            this.State.persianAstro.day = perscal[2];
+            this.State.persianAstro.weekday = this.gWeekDayToPersian(weekday);
+            this.State.persianAstro.leap = this.NormLeap[this.leap_persiana(perscal[0]) ? 1 : 0];
         }
         //  Update Gregorian serial number
         // ---------------------------------------------------------------------------
-        if (this.ON.gregserial.day !== null) {
-            this.ON.gregserial.day = j - this.J0000;
+        if (this.State.gregserial.day !== null) {
+            this.State.gregserial.day = j - this.J0000;
         }
 
 
@@ -452,7 +452,7 @@ class Algorithms {
         // ---------------------------------------------------------------------------
         utime = (j - this.J1970) * (60 * 60 * 24 * 1000);
 
-        this.ON.unixtime = Math.round(utime / 1000);
+        this.State.unixtime = Math.round(utime / 1000);
     }
 
 
@@ -462,25 +462,25 @@ class Algorithms {
      */
     calcGregorian (dateArray) {
         if (dateArray[0] || dateArray[0] === 0) {
-            this.ON.gregorian.year = dateArray[0];
+            this.State.gregorian.year = dateArray[0];
         }
         if (dateArray[1] || dateArray[1] === 0) {
-            this.ON.gregorian.month = dateArray[1];
+            this.State.gregorian.month = dateArray[1];
         }
         if (dateArray[2] || dateArray[2] === 0) {
-            this.ON.gregorian.day = dateArray[2];
+            this.State.gregorian.day = dateArray[2];
         }
         if (dateArray[3] || dateArray[3] === 0) {
-            this.ON.gregorian.hour = dateArray[3];
+            this.State.gregorian.hour = dateArray[3];
         }
         if (dateArray[4] || dateArray[4] === 0) {
-          this.ON.gregorian.minute = dateArray[4];
+          this.State.gregorian.minute = dateArray[4];
         }
         if (dateArray[5] || dateArray[5] === 0) {
-            this.ON.gregorian.second = dateArray[5];
+            this.State.gregorian.second = dateArray[5];
         }
         if (dateArray[6] || dateArray[6] === 0) {
-            this.ON.gregorian.millisecond = dateArray[6];
+            this.State.gregorian.millisecond = dateArray[6];
         }
         this.updateFromGregorian();
     }
@@ -490,14 +490,14 @@ class Algorithms {
      */
     calcJulian () {
         let j, date;
-        j = this.ON.julianday;
+        j = this.State.julianday;
         date = this.jd_to_gregorian(j);
-        this.ON.gregorian.year = date[0];
-        this.ON.gregorian.month = date[1] - 1;
-        this.ON.gregorian.day = date[2];
-        //        this.ON.gregorian.hour = this.pad(time[0], 2, " ");
-        //        this.ON.gregorian.minute = this.pad(time[1], 2, "0");
-        //        this.ON.gregorian.second = this.pad(time[2], 2, "0");
+        this.State.gregorian.year = date[0];
+        this.State.gregorian.month = date[1] - 1;
+        this.State.gregorian.day = date[2];
+        //        this.State.gregorian.hour = this.pad(time[0], 2, " ");
+        //        this.State.gregorian.minute = this.pad(time[1], 2, "0");
+        //        this.State.gregorian.second = this.pad(time[2], 2, "0");
         this.updateFromGregorian();
     }
 
@@ -507,7 +507,7 @@ class Algorithms {
      * @param {*} j
      */
     setJulian (j) {
-        this.ON.julianday = j;
+        this.State.julianday = j;
         this.calcJulian();
     }
 
@@ -518,32 +518,32 @@ class Algorithms {
      */
     calcPersian (dateArray) {
         if (dateArray[0] || dateArray[0] === 0) {
-            this.ON.persian.year = dateArray[0];
+            this.State.persian.year = dateArray[0];
         }
         if (dateArray[1] || dateArray[1] === 0) {
-            this.ON.persian.month = dateArray[1];
+            this.State.persian.month = dateArray[1];
         }
         if (dateArray[2] || dateArray[2] === 0) {
-            this.ON.persian.day = dateArray[2];
+            this.State.persian.day = dateArray[2];
         }
         if (dateArray[3] || dateArray[3] === 0) {
-            this.ON.gregorian.hour = dateArray[3];
+            this.State.gregorian.hour = dateArray[3];
         }
         if (dateArray[4] || dateArray[4] === 0) {
-            this.ON.gregorian.minute = dateArray[4];
+            this.State.gregorian.minute = dateArray[4];
         }
         if (dateArray[5] || dateArray[5] === 0) {
-            this.ON.gregorian.second = dateArray[5];
+            this.State.gregorian.second = dateArray[5];
         }
         if (dateArray[6] || dateArray[6] === 0) {
-            this.ON.gregorian.millisecond = dateArray[6];
+            this.State.gregorian.millisecond = dateArray[6];
         }
 
         this.setJulian(
           this.persian_to_jd(
-            this.ON.persian.year,
-            this.ON.persian.month,
-            this.ON.persian.day)
+            this.State.persian.year,
+            this.State.persian.month,
+            this.State.persian.day)
         );
     }
 
@@ -554,32 +554,32 @@ class Algorithms {
      */
     calcPersiana (dateArray) {
         if (dateArray[0] || dateArray[0] === 0) {
-            this.ON.persianAstro.year = dateArray[0];
+            this.State.persianAstro.year = dateArray[0];
         }
         if (dateArray[1] || dateArray[1] === 0) {
-            this.ON.persianAstro.month = dateArray[1];
+            this.State.persianAstro.month = dateArray[1];
         }
         if (dateArray[2] || dateArray[2] === 0) {
-            this.ON.persianAstro.day = dateArray[2];
+            this.State.persianAstro.day = dateArray[2];
         }
 
         if (dateArray[3] || dateArray[3] === 0) {
-            this.ON.gregorian.hour = dateArray[3];
+            this.State.gregorian.hour = dateArray[3];
         }
         if (dateArray[4] || dateArray[4] === 0) {
-            this.ON.gregorian.minute = dateArray[4];
+            this.State.gregorian.minute = dateArray[4];
         }
         if (dateArray[5] || dateArray[5] === 0) {
-            this.ON.gregorian.second = dateArray[5];
+            this.State.gregorian.second = dateArray[5];
         }
         if (dateArray[6] || dateArray[6] === 0) {
-            this.ON.gregorian.millisecond = dateArray[6];
+            this.State.gregorian.millisecond = dateArray[6];
         }
         this.setJulian(
           this.persiana_to_jd(
-            this.ON.persianAstro.year,
-            this.ON.persianAstro.month,
-            this.ON.persianAstro.day + 0.5)
+            this.State.persianAstro.year,
+            this.State.persianAstro.month,
+            this.State.persianAstro.day + 0.5)
         );
     }
 }