+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix [53](https://github.com/babakhani/PersianDate/issues/53)
 - Fix [56](https://github.com/babakhani/PersianDate/issues/56)
 - Support zero and negative values in input array
+- Prevent ngative value for month/date/hour/minute/seconds   
+-  Fix state manager naming "ON" to "State"
 
 ## [1.0.5] - 2018-2-11
 - Fix [35](https://github.com/babakhani/PersianDate/issues/35)